pinn: move modules to a folder
