+++ fealpy/pinn/modules/__init__.py
@@ -1,0 +1,3 @@
+
+from .module import TensorMapping, Solution
+from .attention import GradAttention
